calculate zeroth order volume moments
No longer hardwired to return 1.0.  This passes all test cases except
for some poorly-coded cases in test_weights.C.  These need rewriting to
use real fvMeshes.

advectionFoamHighOrderFit is crashing because some cell volumes are zero
(some others are negative).  Clearly this is wrong.  Are we trying to
calculate volumes for things that are not really cells (e.g. boundary
faces)?